pyproject.toml
[tool.poetry]
name = "python-web-06"
version = "0.1.0"
description = ""
authors = ["BogdanKozlov35 <bkozlov35@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
faker = "^25.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
